fix double back button and set page title
* Don't pushState everytime initialize is called. Do so only when entity
has been loaded.
* Initialize now monitors entity instead of href. this allows us to set
the document title in the initialize function (we couldn't do that
before since href having a value doesn't mean this.title has a value)
* If no previous state exists (the sequence viewer is loading for the
first time) use replaceState instead of pushState.
* in popSatetListener, if the popped state is the same as the current
activity, go back on more time so that the user is navigated to the
previous activity.